feat: the breakdown into modules has been carried out
