+++ index.html
@@ -231,6 +231,6 @@
   <template id="messages">
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
-  <script src="C:\Users\drvel\2241727-kekstagram-2\js\main.js"></script>
+  <script src="C:\Users\drvel\2241727-kekstagram-2\js\main.js" type="module"></script>
 </body>
 </html>
